(BSR)[API] chore: refactor test to factorise local storage management in tests
