+++ api/tests/test_utils.py
@@ -1,6 +1,8 @@
 import datetime
 import enum
 import operator
+import os
+import pathlib
 import uuid
 
 from pcapi.core.offerers import factories as offerers_factories
@@ -29,3 +31,21 @@ def run_command(app, command_name, *args):
     runner = app.test_cli_runner()
     args = (command_name, *args)
     return runner.invoke(args=args)
+
+
+class StorageFolderManager:
+    storage_folder: pathlib.Path
+
+    def teardown_method(self) -> None:
+        """clear extracts after each tests"""
+        try:
+            for child in self.storage_folder.iterdir():
+                if not child.is_file():
+                    continue
+                child.unlink()
+        except FileNotFoundError:
+            pass
+
+    def setup_method(self):
+        """Create the folder to work with"""
+        os.makedirs(self.storage_folder, exist_ok=True)
